+++ packages/e2eos-angular-antd/src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { GridPageComponent } from './pages/grid-page/grid-page.component';
 import { InputPageComponent } from './pages/input-page/input-page.component';
 import {LayoutPageComponent} from './pages/layout-page/layout-page.component';
 import {CardPageComponent} from './pages/card-page/card-page.component';
+import {CollapsePageComponent} from './pages/collapse-page/collapse-page.component';
 
 const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: 'button' },
@@ -14,7 +15,8 @@ const routes: Routes = [
   { path: 'input', component: InputPageComponent },
   { path: 'grid', component: GridPageComponent },
   { path: 'layout', component: LayoutPageComponent },
-  { path: 'card', component: CardPageComponent }
+  { path: 'card', component: CardPageComponent },
+  { path: 'collapse', component: CollapsePageComponent }
 ];
 
 @NgModule({
