+++ packages/e2eos-angular-antd/src/app/app.component.html
@@ -43,6 +43,11 @@
             <a routerLink="/card">Card 卡片</a>
           </li>
         </ul>
+        <ul>
+          <li nz-menu-item nzMatchRouter>
+            <a routerLink="/collapse">Collapse 折叠面板</a>
+          </li>
+        </ul>
       </li>
       <li nz-submenu nzOpen nzTitle="反馈" nzIcon="message">
       </li>
