--- packages/e2eos-angular-antd/src/app/components/collapse-panel/collapse-panel.component.html
@@ -1,3 +1,0 @@
-<nz-collapse><nz-collapse-panel [nzDisabled]="attrs.nzDisabled" [nzHeader]="attrs.nzHeader" [nzExpandedIcon]="attrs.nzExpandedIcon" [nzExtra]="attrs.nzExtra" [nzShowArrow]="attrs.nzShowArrow" [nzActive]="attrs.nzActive" >
-  <e2eos-ng-antd-container [node]="n" *ngFor="let n of node.children"></e2eos-ng-antd-container>
-</nz-collapse-panel></nz-collapse>
