+++ packages/e2eos-angular-antd/src/app/pages/collapse-page/collapse-page.component.html
@@ -1,0 +1,3 @@
+<e2eos-ng-antd-json-editor [(data)]="node"></e2eos-ng-antd-json-editor>
+
+<e2eos-ng-antd-container [node]="node"></e2eos-ng-antd-container>
